requirements.txt
pywin32
keyboard
pyperclip
dearpygui
pyspellchecker
beautifulsoup4
